requirements.txt
Flask==0.10.1
Flask-WTF==0.11
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.2
Werkzeug==0.10.1
itsdangerous==0.24
Flask-SQLAlchemy==2.0
MySQL-python==1.2.5
